update dev docker and JWT install
