+++ app/routers/datasets.py
@@ -1,17 +1,20 @@
+import os
 import datetime
-from typing import List
+from typing import List, Optional
 import json
 from bson import ObjectId
-from fastapi import APIRouter, Request, HTTPException, Depends
+from fastapi import APIRouter, Request, HTTPException, Depends, File, UploadFile
 from pymongo import MongoClient
 from pydantic import Json
+from minio import Minio
 from fastapi.encoders import jsonable_encoder
+
 from app import dependencies
 from app.models.datasets import Dataset
-from app.models.files import ClowderFile
+from app.models.files import ClowderFile, FileVersion
 from app.auth import AuthHandler
-import os
-from minio import Minio
+from app.config import settings
+
 
 router = APIRouter()
 
@@ -128,3 +131,48 @@ async def delete_dataset(
         res = await db["datasets"].delete_one({"_id": ObjectId(dataset_id)})
         return {"deleted": dataset_id}
     raise HTTPException(status_code=404, detail=f"Dataset {dataset_id} not found")
+
+
+@router.post("/{dataset_id}/files", response_model=ClowderFile)
+async def save_file(
+    dataset_id: str,
+    user_id=Depends(auth_handler.auth_wrapper),
+    db: MongoClient = Depends(dependencies.get_db),
+    fs: Minio = Depends(dependencies.get_fs),
+    file: UploadFile = File(...),
+    file_info: Optional[ClowderFile] = None,
+):
+    if (
+        dataset := await db["datasets"].find_one({"_id": ObjectId(dataset_id)})
+    ) is not None:
+        # Prepare new file entry
+        f = dict(file_info) if file_info is not None else {}
+        user = await db["users"].find_one({"_id": ObjectId(user_id)})
+        f["name"] = file.filename
+        f["version"] = FileVersion(creator=user["_id"])
+        f["views"] = 0
+        f["downloads"] = 0
+
+        # Add to db and update dataset
+        new_file = await db["files"].insert_one(f)
+        found = await db["files"].find_one({"_id": new_file.inserted_id})
+        new_file_id = found["_id"]
+        updated_dataset = await db["datasets"].update_one(
+            {"_id": ObjectId(dataset_id)}, {"$push": {"files": ObjectId(new_file_id)}}
+        )
+
+        # Use unique ID as key for file storage
+        while content := file.file.read(
+            settings.MINIO_UPLOAD_CHUNK_SIZE
+        ):  # async read chunk
+            fs.put_object(
+                settings.MINIO_BUCKET_NAME,
+                str(new_file_id),
+                io.BytesIO(content),
+                length=-1,
+                part_size=settings.MINIO_UPLOAD_CHUNK_SIZE,
+            )  # async write chunk to minio
+
+        return ClowderFile.from_mongo(found)
+    else:
+        raise HTTPException(status_code=404, detail=f"Dataset {dataset_id} not found")
