+++ app/routers/files.py
@@ -18,54 +18,13 @@ from pymongo import MongoClient
 from minio import Minio
 
 from app import dependencies
-from app.models.files import ClowderFile
+from app.models.files import ClowderFile, FileVersion
 from app.auth import AuthHandler
 from app.config import settings
 
 router = APIRouter()
 
 auth_handler = AuthHandler()
-
-
-@router.post("/{dataset_id}", response_model=ClowderFile)
-async def save_file(
-    dataset_id: str,
-    user_id=Depends(auth_handler.auth_wrapper),
-    db: MongoClient = Depends(dependencies.get_db),
-    fs: Minio = Depends(dependencies.get_fs),
-    file: UploadFile = File(...),
-    file_info: Optional[ClowderFile] = None,
-):
-    # First, add to database and get unique ID
-    f = dict(file_info) if file_info is not None else {}
-    user = await db["users"].find_one({"_id": ObjectId(user_id)})
-    dataset = await db["datasets"].find_one({"_id": ObjectId(dataset_id)})
-    f["name"] = file.filename
-    f["creator"] = user["_id"]
-    f["views"] = 0
-    f["downloads"] = 0
-    new_file = await db["files"].insert_one(f)
-    found = await db["files"].find_one({"_id": new_file.inserted_id})
-    new_file_id = found["_id"]
-
-    # TODO: If file with same filename exists in dataset, default to overwriting (?)
-    updated_dataset = await db["datasets"].update_one(
-        {"_id": ObjectId(dataset_id)}, {"$push": {"files": ObjectId(new_file_id)}}
-    )
-
-    # Second, use unique ID as key for file storage
-    while content := file.file.read(
-        settings.MINIO_UPLOAD_CHUNK_SIZE
-    ):  # async read chunk
-        fs.put_object(
-            settings.MINIO_BUCKET_NAME,
-            str(new_file_id),
-            io.BytesIO(content),
-            length=-1,
-            part_size=settings.MINIO_UPLOAD_CHUNK_SIZE,
-        )  # async write chunk to minio
-
-    return ClowderFile.from_mongo(found)
 
 
 @router.put("/{file_id}")
@@ -121,6 +80,8 @@ async def download_file(
             {"_id": ObjectId(file_id)}, {"$inc": {"downloads": 1}}
         )
         return response
+    else:
+        raise HTTPException(status_code=404, detail=f"File {file_id} not found")
 
 
 @router.delete("/{file_id}")
